+++ applications_user/protopirate/scenes/protopirate_scene_receiver_config.c
@@ -255,8 +255,8 @@ static void protopirate_scene_receiver_config_set_rx_profile(VariableItem* item)
         text,
         sizeof(text),
         "%lu.%02lu",
-        profile->frequency / 1000000,
-        (profile->frequency % 1000000) / 10000);
+        (unsigned long)(profile->frequency / 1000000),
+        (unsigned long)((profile->frequency % 1000000) / 10000));
     variable_item_set_current_value_text(frequency, text);
     variable_item_set_current_value_index(
         frequency, protopirate_scene_receiver_config_next_frequency(profile->frequency, app));
